Move mixins folder
