+++ src/components/article/Image.js
@@ -1,7 +1,7 @@
 'use strict'
 import _ from 'lodash'
 import classNames from 'classnames'
-import FitwidthMixin from '../mixins/FitwidthMixin'
+import FitwidthMixin from './mixins/FitwidthMixin'
 import { getScreenType } from '../../lib/screen-type'
 import BlockAlignmentWrapper from './BlockAlignmentWrapper'
 import React, { Component } from 'react'
