+++ src/components/article/mixins/FitwidthMixin.js
